Cleanup + Add better exception handling
