+++ Postgrest/Attributes/ReferenceAttribute.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Runtime.CompilerServices;
-using Newtonsoft.Json;
 using Postgrest.Extensions;
 using Postgrest.Models;
