+++ Postgrest/Constants.cs
@@ -1,5 +1,4 @@
-﻿using Postgrest.Attributes;
-using Supabase.Core.Attributes;
+﻿using Supabase.Core.Attributes;
 
 namespace Postgrest
 {
