+++ Postgrest/Extensions/TypeExtensions.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Text;
 
 namespace Postgrest.Extensions
 {
