+++ Postgrest/Linq/SelectExpressionVisitor.cs
@@ -1,11 +1,7 @@
 ﻿using Postgrest.Attributes;
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Linq.Expressions;
-using System.Reflection;
-using System.Text;
-using static Postgrest.Constants;
 
 namespace Postgrest.Linq
 {
