+++ Postgrest/Linq/SetExpressionVisitor.cs
@@ -1,12 +1,7 @@
-﻿using Newtonsoft.Json.Linq;
-using Postgrest.Attributes;
+﻿using Postgrest.Attributes;
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Linq.Expressions;
-using System.Reflection;
-using System.Text;
-using static Postgrest.Constants;
 
 namespace Postgrest.Linq
 {
