+++ Postgrest/QueryOptions.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using Postgrest.Attributes;
 using Postgrest.Extensions;
 using Supabase.Core.Attributes;
