+++ PostgrestTests/ExceptionTests.cs
@@ -1,0 +1,6 @@
+﻿namespace PostgrestTests
+{
+	internal class ExceptionTests
+	{
+	}
+}
